[NOTIFICATIONS][item]:notification item as a component for menu lisibility
