+++ src/shared/navbar/notifications/NotificationMenu.tsx
@@ -1,6 +1,7 @@
 import { Box, Menu, Typography, styled } from "@mui/material";
 import { JSX } from "react";
 import { NotificationItem } from "./types";
+import { NotificationListItem } from "./NotificationListItem";
 
 type NotificationMenuProps = {
   anchorEl: HTMLElement | undefined;
@@ -30,31 +31,9 @@ const MenuContent = styled(Box)(({ theme }) => ({
   padding: theme.spacing(1.5),
 }));
 
-const NotificationEntry = styled(Box, {
-  shouldForwardProp: (prop) => prop !== "read",
-})<{ read: boolean }>(({ theme, read }) => ({
-  padding: theme.spacing(1, 0),
-  borderBottom: `1px solid ${theme.palette.gray[500]}`,
-  cursor: "pointer",
-  opacity: read ? 0.50 : 1,
-  "&:last-of-type": {
-    borderBottom: "none",
-  },
-}));
-
 const EmptyState = styled(Typography)(({ theme }) => ({
   color: theme.palette.gray[300],
   padding: theme.spacing(1, 0),
-}));
-
-const NonImportantTypography = styled(Typography)(({ theme }) => ({
-  color: theme.palette.gray[300],
-}));
-
-const ActionsRow = styled(Box)(({ theme }) => ({
-  display: "flex",
-  gap: theme.spacing(1),
-  marginTop: theme.spacing(0.5),
 }));
 
 export const NotificationMenu = ({
@@ -72,29 +51,11 @@ export const NotificationMenu = ({
           <EmptyState variant="body2">No notifications yet.</EmptyState>
         ) : (
           notifications.map((notification) => (
-            <NotificationEntry
+            <NotificationListItem
               key={notification.id}
-              read={notification.read}
-            >
-              <Typography variant="body2" fontWeight={600}>
-                {notification.title || notification.type}
-              </Typography>
-              <Typography variant="caption">
-                {notification.message}
-              </Typography>
-              <NonImportantTypography variant="caption" display="block">
-                {new Date(notification.createdAt).toLocaleString()}
-              </NonImportantTypography>
-              <ActionsRow>
-                <Typography
-                  variant="caption"
-                  sx={{ cursor: "pointer", textDecoration: "underline" }}
-                  onClick={() => onMarkAsRead(notification.id)}
-                >
-                  Mark as read
-                </Typography>
-              </ActionsRow>
-            </NotificationEntry>
+              notification={notification}
+              onMarkAsRead={onMarkAsRead}
+            />
           ))
         )}
       </MenuContent>
